+++ ChangeLog.md
@@ -4,7 +4,8 @@ Blink Framework Change Log
 0.1.1 xx xx, 2015
 ---------------------
 
-
+- Enh: Added file uploading support
+- Enh: Added CgiServer for php-fpm or Apache's mod_php support
 
 0.1.0 October 19, 2015
 ---------------------